src/requirements.txt
coverage==3.6
dj-database-url==0.5.0
Django==2.1.4
django-environ==0.4.5
django-redis==4.10.0
djangorestframework==3.9.0
gunicorn==19.9.0
psycopg2-binary==2.7.6.1
python-memcached==1.59
pytz==2018.7
redis==3.0.1
six==1.12.0
uWSGI==2.0.17.1
